feat(core): Make Core portable
