+++ packages/core-bridge/index.js
@@ -1,0 +1,12 @@
+const { getPrebuiltPath, PrebuildError } = require('./common');
+
+try {
+  const prebuiltPath = getPrebuiltPath();
+  module.exports = require(prebuiltPath);
+} catch (err) {
+  if (err instanceof PrebuildError) {
+    module.exports = require('./default-build/index.node');
+  } else {
+    throw err;
+  }
+}
