+++ packages/core-bridge/scripts/build.js
@@ -1,26 +1,12 @@
 const path = require('path');
 const arg = require('arg');
-const os = require('os');
 const fs = require('fs');
 const which = require('which');
 const { spawnSync } = require('child_process');
 const { version } = require('../package.json');
+const { targets, getPrebuiltPath, PrebuildError } = require('../common');
 
 process.chdir(path.resolve(__dirname, '..'));
-
-// List of tested compile targets
-const targets = [
-  'x86_64-apple-darwin',
-  'aarch64-apple-darwin',
-  'x86_64-unknown-linux-gnu',
-  'aarch64-unknown-linux-gnu',
-  // TODO: this is not supported on macos
-  'x86_64-pc-windows-msvc',
-  'x86_64-pc-windows-gnu',
-];
-
-const archAlias = { x64: 'x86_64', arm64: 'aarch64' };
-const platformMapping = { darwin: 'apple-darwin', linux: 'unknown-linux-gnu', win32: 'pc-windows-gnu' };
 
 const args = arg({
   '--help': Boolean,
@@ -71,7 +57,7 @@ const buildRelease = args['--release'] || process.env.BUILD_CORE_RELEASE !== und
 function compile(target) {
   console.log('Compiling bridge', { target, buildRelease });
 
-  const out = target ? `releases/${target}/index.node` : 'index.node';
+  const out = target ? `releases/${target}/index.node` : 'default-build/index.node';
   try {
     fs.unlinkSync(out);
   } catch (err) {
@@ -103,35 +89,6 @@ function compile(target) {
   }
 }
 
-class PrebuildError extends Error {
-  constructor(message) {
-    super(message);
-    this.name = 'PrebuildError';
-  }
-}
-
-function usePrebuilt() {
-  const arch = archAlias[os.arch()];
-  if (arch === undefined) {
-    throw new PrebuildError(`No prebuilt module for arch ${os.arch()}`);
-  }
-  const platform = platformMapping[os.platform()];
-  if (platform === undefined) {
-    throw new PrebuildError(`No prebuilt module for platform ${os.platform()}`);
-  }
-  const source = path.resolve(__dirname, '../releases', `${arch}-${platform}`, 'index.node');
-  const target = path.resolve(__dirname, '..', 'index.node');
-  try {
-    fs.copyFileSync(source, target);
-    console.log('Copied prebuilt bridge module', { source, target });
-  } catch (err) {
-    if (err.code === 'ENOENT') {
-      throw new PrebuildError(`No prebuilt module found at ${source}`);
-    }
-    throw err;
-  }
-}
-
 if (requestedTargets.length > 0) {
   // NOTE: no forceBuild
   for (const target of requestedTargets) {
@@ -140,9 +97,10 @@ if (requestedTargets.length > 0) {
 } else {
   if (!forceBuild) {
     try {
-      usePrebuilt();
+      getPrebuiltPath();
     } catch (err) {
       if (err instanceof PrebuildError) {
+        console.warn(err.message);
         compile();
       } else {
         throw err;
